refactor/extract: moves KeyRotation in preparation for renaming
